+++ src/talispam/core.clj
@@ -8,55 +8,16 @@
          '[talispam.filter :as f]
          '[talispam.utils :as utils]
          '[talispam.whitelist :as w]
-         '[clojure.tools.cli :refer [parse-opts]]
+         '[cli-matic.core :refer [run-cmd]]
          '[com.stuartsierra.frequencies :as freq]
          '[tlight.spin :refer [spin done]])
-
-(def cli-options
-  [["-m" "--mbox FILE" "mailbox to analyze"
-    :validate [#(utils/is-mbox? %) "must be a mbox file"]]
-   ["-h" "--help"]])
-
-(defn- usage [options-summary]
-  (->> [(str c/program-name " " c/program-version)
-        ""
-        "Usage: talispam [action [options]]"
-        ""
-        "Options:"
-        options-summary
-        ""
-        "Actions:"
-        "  learn        train talispam classifier"
-        "  score        print ham/spam score for stdin"
-        "  whitelist    print a list of addresses in ham corpus"
-        "  print-db     print all words from classifier db by spam score"
-        "  stats        print stats summary for a mbox"]
-       (s/join \newline)))
-
-(defn- error-msg [errors]
-  (str "The following errors occurred while parsing your command:\n\n"
-       (s/join \newline errors)))
-
-(defn- validate-args
-  [args]
-  (let [{:keys [options arguments errors summary]} (parse-opts args cli-options)]
-    (cond
-      (:help options)
-      {:exit-message (usage summary) :ok? true}
-      errors
-      {:exit-message (error-msg errors)}
-      (and (= 1 (count arguments))
-           (#{"learn" "score" "whitelist" "print-db" "stats"} (first arguments)))
-      {:action (first arguments) :options options}
-      :else
-      {:action nil})))
 
 (defn- exit [status msg]
   (println msg)
   (System/exit status))
 
 ;; train classifier
-(defn- learn! []
+(defn- learn! [parms]
   (spin :type :spin1 :ms 200)
   (print "building classifier db ")
   (f/learn)
@@ -70,7 +31,7 @@
 
 (defn- load-db []
   (if (not (db/exists-db))
-    (exit 1 "classifier db not found, see talispam -h"))
+    (exit 1 (str "classifier db not found, see " c/program-name " --help")))
   (try
     (db/load-db)
     (catch Exception e (exit 1 "error loading classifier db"))))
@@ -89,7 +50,10 @@
                                   (if (:use (:whitelist @c/config))
                                     (w/whitelisted? (utils/get-sender in))))))))
 
-(defn- print-db []
+(defn- classify-score [parms]
+  (classify *in* 'score))
+
+(defn- print-db [parms]
   (load-db)
   (println
    (s/join \newline
@@ -106,6 +70,32 @@
     (doseq [[k v] (map vector (keys res) (vals res))]
       (println (str (name k) " " v)))))
 
+;; cli-matic config
+(def CONFIGURATION
+  {:app         {:command     c/program-name
+                 :description c/program-description
+                 :version     c/program-version}
+   :commands    [{:command     "learn"
+                  :description ["train talispam classifier"]
+                  :opts []
+                  :runs        learn!}
+                 {:command     "score"
+                  :description ["print ham/spam score for stdin"]
+                  :opts []
+                  :runs        classify-score}
+                 {:command     "whitelist"
+                  :description ["print a list of addresses in ham corpus"]
+                  :opts []
+                  :runs        w/print-whitelist-from-corpus}
+                 {:command     "print-db"
+                  :description ["print all words from classifier db by spam score"]
+                  :opts []
+                  :runs        print-db}
+                 {:command     "stats"
+                  :description ["print stats summary for a mbox"]
+                  :opts        [{:option "mbox" :short "m" :default :present :as "mbox file" :type :string}]
+                  :runs        stats}]})
+
 (defn -main [& args]
   ;; load configuration
   (try
@@ -117,14 +107,7 @@
     (try
       (dict/load-dictionary!)
       (catch Exception e (exit 1 (.getMessage e)))))
-  
-  (let [{:keys [action options exit-message ok?]} (validate-args args)]
-    (if exit-message
-      (exit (if ok? 0 1) exit-message)
-      (case action
-        "learn" (learn!)
-        "score" (classify *in* 'score)
-        "whitelist" (w/print-whitelist-from-corpus)
-        "print-db" (print-db)
-        "stats" (stats options)
-        nil (classify *in*)))))
+
+  (if (nil? args)
+    (classify *in*)
+    (run-cmd args CONFIGURATION)))